car_rental main CRUD done

+++ carRental/src/main/java/com/example/carRental/controller/CarController.java
@@ -5,7 +5,9 @@ import com.example.carRental.dto.CarRequestDTO;
 import com.example.carRental.dto.CarResponseDTO;
 import com.example.carRental.model.Car;
 import com.example.carRental.service.CarService;
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
@@ -18,8 +20,6 @@ import java.util.List;
 public class CarController {
 
   private final CarService carService;
-  Car car = new Car();
-  List<Car> cars = new ArrayList<>();
 
   @Autowired
   public CarController(CarService carService) {
@@ -31,10 +31,11 @@ public class CarController {
     return ResponseEntity.ok(CarMapper.toCarResponseDTOList(carService.findAllCars()));
   }
 
-//  @GetMapping("/cars/available")
-//  public ResponseEntity<List<CarResponseDTO>> getCarsAvailable() {
-//    if ()
-//  }
+  @GetMapping("/cars/available")
+  public ResponseEntity<List<CarResponseDTO>> getCarsAvailable() {
+    List<Car> carsAvailable = carService.findAllCars().stream().filter(car -> car.getStatus().equals("AVAILABLE")).toList();
+    return ResponseEntity.ok(CarMapper.toCarResponseDTOList(carsAvailable));
+  }
 
   @PostMapping("/cars")
   public ResponseEntity<?> addCar(@RequestBody CarRequestDTO carRequestDTO) {
@@ -44,10 +45,52 @@ public class CarController {
     return ResponseEntity.created(ServletUriComponentsBuilder.fromCurrentRequest().path("/{id}").buildAndExpand(savedCar.getId()).toUri()).body(CarMapper.toCarResponseDTO(savedCar));
   }
 
-  @DeleteMapping("/cars/{id}")
-  public ResponseEntity<Void> deleteCar(@PathVariable long id) {
+  @PutMapping("/cars/{id}")
+  public ResponseEntity<?> updateCar(@PathVariable long id, @RequestBody CarRequestDTO carRequestDTO) {
+
+    if (carService.existsCarById(id)) {
+
+      Car carFromDb = carService.findCarById(id).get();
+
+      CarMapper.updateCarFromDTO(carRequestDTO, carFromDb);
+      carService.saveCar(carFromDb);
+
+      return ResponseEntity.ok(carRequestDTO);
+    }
+
+    Car savedCar = carService.savedCar(CarMapper.toCar(carRequestDTO));
+
+    return ResponseEntity.created(ServletUriComponentsBuilder.fromCurrentRequest()
+                    .replacePath("/api/cars/{id}")
+                    .buildAndExpand(savedCar.getId())
+                    .toUri())
+            .body(savedCar);
+  }
+
+  @PutMapping("/cars/{id}/rent")
+  public ResponseEntity<?> changeCarStatus(@PathVariable long id) {
+
     if (!carService.existsCarById(id)) {
-      return ResponseEntity.notFound().build();
+      return ResponseEntity.status(HttpStatus.NOT_FOUND).body("car was not found");
+    }
+    Car carFromDb = carService.findCarById(id).get();
+    carFromDb.setStatus(carFromDb.getStatus().equals("AVAILABLE") ? "RENTED" : "AVAILABLE");
+    carService.saveCar(carFromDb);
+
+
+    return ResponseEntity.ok(CarMapper.toCarResponseDTO(carFromDb));
+  }
+
+
+  @DeleteMapping("/cars/{id}")
+  public ResponseEntity<?> deleteCar(@PathVariable long id) {
+
+    if (!carService.existsCarById(id)) {
+      return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Car not found.");
+    }
+
+    if (carService.findById(id).getStatus().equals("RENTED")) {
+      return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Car is rented. Can't remove that one");
     }
     carService.deleteCarById(id);
     return ResponseEntity.noContent().build();
